Move MainWindow to ui.window package
